Fixed "overlay" issue not sizing correctly
The last way I was trying to overlay one div on another caused problems
(I think) with bootstrap, and it made things nearly impossible. I was
totally trying to do things the hard way, because what I've changed it
to do is hide the main content div and replace it with the one I want to
show. Makes things a LOT easier.